// resources/js/api/generated/models/container.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { ContainerLabels } from './containerLabels';
import type { ContainerMount } from './containerMount';
import type { ContainerNetwork } from './containerNetwork';
import type { HealthStatus } from './healthStatus';
import type { Port } from './port';
import type { ResourceLimits } from './resourceLimits';
import type { RestartPolicy } from './restartPolicy';

export interface Container {
  command?: string[];
  created?: string;
  /** @nullable */
  exit_code?: number | null;
  finished?: string;
  health?: HealthStatus | null;
  image: string;
  labels?: ContainerLabels;
  mounts?: ContainerMount[];
  name: string;
  networks?: ContainerNetwork[];
  ports?: Port[];
  resource_limits?: ResourceLimits | null;
  restart_policy?: RestartPolicy | null;
  started?: string;
  state: string;
  user?: string;
  working_dir?: string;
}